First commit.
This commit brings the implementation of the infoGail circle shown in the InfoGAIL: Interpretable Imitation Learning from Visual Demonstrations paper as well as the Decision Diffuser brought by the Is Conditional Generative Modeling all you need for Decision-Making? paper. The idea here is to in the future use the decision the diffuser to solve the circle environment defined by the InfoGail paper.